Improve and test OIDC discovery (#14)
- Make OIDC discovery actually use proper endpoints
- Introduce tests for OIDC discovery
- Introduce test to assert when OIDC discovery should happen

At first, I wanted to use the Url from the `url` crate in the builder,
since that would make it very clear what's expected. However, if I
entered "https://authorization-server.com" as issuer_uri, when running
to_string() or as_ref() it'd return "https://authorization-server.com/".
The trailing slash is a problem since I use this property to also find
out about the expected `iss` claim, which need to be an exact match.