Set MSRV to 1.45.0 (#64)
In the last commit I updated version of salvo which now required
rust-version 1.85.0 and edition 2024. The commit resulted in
tower-oauth2-resource-server having the same requirements.

In this PR I set edition 2024 on the salvo example crate only.
I've also set the MSRV 1.45.0 (edition 2018) - I believe this is the
lowest I can support (usage of strip_prefix in jwt_extract became stable
in that version).